MultiHeadAttn accepts out_dim parameter
The MultiHeadAttn layer initializer accepts an additional parameter
specifying the dimension of the outputs.

This will not change its usage in the transformer blocks. Because of the
residual connections the output dimension MUST match the input
dimension. However, when used as a standalone layer, this modification
allows for more flexibility.